add Dockerfile to build an image with delphy
It would be helpful to have delphy available in a container image, for execution via WDL+Terra.bio, or via other pipeline systems that rely on Docker to encapsulate software and required dependencies.

This adds a Dockerfile for building a small image containing a compiled delphy binary.

The Docker build takes part in several stages, with a separate cacheable image for each:
1) setup of a basic build environment, with conda (micromamba) for dependency management (also adds a yaml file defining the conda environment)
2) installation of linked dependencies, via conan and by updating git submodules (depending on how the dependencies are specified), followed by compilation of a static binary
3) copying of the static binary from the compilation stage to a minimal image based on Alpine Linux (which favors musl and lacks glibc, hence the static compilation of delphy)

The resulting distributable image—at present—is approximately 18MB in size, with half of that being the (static) delphy binary itself.

The Dockerfile can be auto-built if this repo is added to quay.io (or another registry that can pull and build from GitHub)