Stop trying to separate ID from comment in FASTA identifier lines
We used to separate FASTA identifier lines into an "ID" and a "comment"
field, delimited by a space, after the description in
https://www.ncbi.nlm.nih.gov/genbank/fastaformat/). This has caused no
end of grief for real users.  For example, an identifier line reading
```
>My Very own Viral Sequence|2025-04-24
```
was read as sequence with ID `"My"`.  Since that lacks a date, Delphy
would complain and the sequence would be ignored.

Both BEAST and IQ-Tree read the entirety of the line after the `>`, trim
it, and treat _that_ as the sequence ID.  We now do the same.